feat: avoid doing multiple RPC requests for files.add, fixes #522
