+++ src/files/add.js
@@ -2,8 +2,8 @@
 
 const isStream = require('is-stream')
 const promisify = require('promisify-es6')
-const DAGNodeStream = require('../utils/dagnode-stream')
 const ProgressStream = require('../utils/progress-stream')
+const ConverterStream = require('../utils/converter')
 
 module.exports = (send) => {
   return promisify((files, opts, callback) => {
@@ -44,9 +44,9 @@ module.exports = (send) => {
 
     const request = { path: 'add', files: files, qs: qs, progress: opts.progress }
 
-    // Transform the response stream to DAGNode values
-    const transform = (res, callback) => DAGNodeStream
-      .streamToValue(send, ProgressStream.fromStream(opts.progress, res), callback)
+    const transform = (res, callback) => {
+      ConverterStream.streamToValue(send, ProgressStream.fromStream(opts.progress, res), callback)
+    }
     send.andTransform(request, transform, callback)
   })
 }
